[DIAGNOSTICS] Fix condition where scheduler task can't be found. (#2878)
This happens due to a name mismatch. The DCOS_SERVICE_NAME label in SDK services and the service name provided through the CLI (i.e. dcos $package --name=$service_name ...) may or may not be prefixed by a slash, so comparing the two as-is sometimes doesn't work.

This change makes sure to get the scheduler task with and without a slash prefix in the name.